# spi: Add SPI block-write to C++ and HAL for performance