--- a/targets/TARGET_Atmel/TARGET_SAM_CortexM4/spi_api.c
+++ b/targets/TARGET_Atmel/TARGET_SAM_CortexM4/spi_api.c
@@ -296,6 +296,20 @@ int  spi_master_write(spi_t *obj, int value)
     return 0;
 }
 
+int spi_master_block_write(spi_t *obj, const char *tx_buffer, int tx_length, char *rx_buffer, int rx_length) {
+    int total = (tx_length > rx_length) ? tx_length : rx_length;
+
+    for (int i = 0; i < total; i++) {
+        char out = (i < tx_length) ? tx_buffer[i] : 0xff;
+        char in = spi_master_write(obj, out);
+        if (i < rx_length) {
+            rx_buffer[i] = in;
+        }
+    }
+
+    return total;
+}
+
 /** Check if a value is available to read
  *
  * @param[in] obj The SPI peripheral to check