… targets/TARGET_NXP/TARGET_LPC81X/spi_api.c
… targets/TARGET_NXP/TARGET_LPC82X/spi_api.c
… targets/TARGET_ONSEMI/TARGET_NCS36510/spi_api.c
… targets/TARGET_RENESAS/TARGET_RZ_A1H/spi_api.c
… targets/TARGET_RENESAS/TARGET_VK_RZ_A1H/spi_api.c
… targets/TARGET_STM/stm_spi_api.c
… targets/TARGET_Silicon_Labs/TARGET_EFM32/spi_api.c
… targets/TARGET_WIZNET/TARGET_W7500x/spi_api.c