This is based on previous work done by @simonqhughes to speed up the SD card implementation, it has a wider impact than just the SD card, so I'm making this pr against the current SPI implementation.

Replaced with assertion on current mutex owner. This requires locking in the caller, but without external locking the current behaviour is still at risk of contention for transactions > 1 byte.

Tested with sd-driver basic tests:

|                                   |  1MHz SPI | 40 MHz SPI |
|-------------|----|----|
| with byte-locking      | 32.911s | 25.386s |
| without byte-locking | 25.386s | 15.977s |

**Note:** This does create the concern that single-byte writes are no longer thread-safe when used across multiple threads. However, there is very few use cases where single-byte SPI transactions are used, and most transactions already have external locking.

cc @c1728p9, @sg- 

---

EDIT:

Sorry about the delay, got dragged into some other issues.

I have more data for you all, there's still quite a few patches I need to put up places but I have a better test that measure block writes vs block reads vs spi traffic (this is all with the SD card). I was also tinkering with the K64F's ~dma~ dspi api and put together a block_write like the one @kjbracey-arm was suggesting that takes advantage of the K64F's ~dma~ dspi api (this was before the talk of using the async api).

Anyways here's what I found with GCC_ARM and K64F at 40MHz:

|    |   SD write speed   |   SD read speed |    SPI speed  |
|---|-----------------------|----------------------|---|
| with byte-writes      | 463.263Kbps      |      792.419Kbps|         913.783Kbps |
| with block-writes    |   757.669Kbps      |      2.338Mbps       |     3.997Mbps |
| with hal block-writes |  878.342Kbps      |      4.239Mbps        |    17.809Mbps |

There is definitely motivation for block writes at the hal level in terms of raw SPI throughput.

As for why we don't see a proportional increase in throughput on the SD card, well I think this image sums it up quite nicely:
![image](https://cloud.githubusercontent.com/assets/1075160/26221504/624a3568-3bdc-11e7-8bc9-53aecc682151.png)

---

EDIT: EDIT:

I've gone ahead and updated this pr with the hal-level block writes, a default implementation in all of the target's spi implementations, and the example block write implementation using ~DMA~ the DSPI API in the k64f (33e9129) that I used for evaluation.

I've gone ahead and updated the PR description and title to reflect this.