@0xc0170 2017-04-21
+1 for the change, however this is a breaking change (should be noted here and even in the commit message), so users updating to the newer version would be aware of this. 

What about the rest of drivers? I2C has the same problem (write for one byte locks/unlocks).