@sg- 2017-04-21
Is the point to then allow external synchronization around CS and override lock/unlock? The performance increase doesn't look much better but I don't know what test was used. Looks like write still calls acquire which still calls write on a per write basis.

Would having a block write addition not provide more performance?