@sg- 2017-04-21
A nit but might consider updating the protected member naming typo (not your addition)? https://github.com/geky/mbed/blob/863717f427cb3859de5bd7091200c52d09f6ebc1/drivers/SPI.h#L263