@geky 2017-04-21
Sorry, I should have clarified, this was tested by timing the `sd-driver-features-tests-filesystem-basic` set of tests, so there is other filesystem operations going on:
https://github.com/ARMmbed/sd-driver/blob/master/features/TESTS/filesystem/basic/basic.cpp

With the locks removed from acquire this is now:

|                                   |  1MHz SPI | 40 MHz SPI |
|-------------|----|----|
| with byte-locking      | 32.911s | 21.264s |
| without byte-locking | 25.386s | 10.816s |

I2C should probably be updated as well for consistency, but it is less priority for the reason @c1728p9 mentioned.

I'll see what benefit there might be for a block write for the K64F, but that would be a much bigger HAL change.