@c1728p9 2017-04-23
@geky I didn't mean test block write at the HAL level. I just meant calling spi_master_write in a loop with a block write function. That way the lock, acquire and class call are needed only once for the block.