@0xc0170 2017-04-24
@sg- and @c1728p9 What do you think about this patch? We shall address this low speed problem. As mentioned already we got some options (were already mentioned here).

> Would having a block write addition not provide more performance?

This would be one - having block write as i2c has that would have lock/unlock.  Or as is it here, lock/unlock is up to an app via lock/unlock methods, they are not invoked within write and acquire.

> I'll see what benefit there might be for a block write for the K64F, but that would be a much bigger HAL change.

Would it? Having write() method that just wraps around single-byte write() with locking?