@geky 2017-04-25
As @c1728p9 pointed out, I was confused on what layer the block write was inteded for.

Here's the measurements for the `sd-driver-features-tests-filesystem-basic` tests with a block write implementation in the SPI class ([here](https://github.com/geky/mbed/commit/e7db61ec04c0440196abf2c034b9945b06c5a7a0)):

|                                   |  1MHz SPI | 40 MHz SPI |
|-------------|----|----|
| with byte-locking      | 32.911s | 21.264s |
| without byte-locking | 25.386s | 10.816s |
| with block-writes | 26.206s | 9.889s |

The block-writes and byte-writes without locking are comparable. There are still some patterns that don't map to block writes, such as [sync patterns](https://github.com/ARMmbed/sd-driver/blob/master/features/filesystem/sd/SDBlockDevice.cpp#L206-L208) and [commands](https://github.com/ARMmbed/sd-driver/blob/master/features/filesystem/sd/SDBlockDevice.cpp#L398-L403), but these are a low percentage of SPI traffic.

---

I personally think taking in both removing the locking from byte-writes and the block-writes would be beneficial to have, so I don't really have a preference. Do we want to go with the block-writes instead?