@0xc0170 2017-04-25
Thanks @geky for testing all options ! 

Having extending SPI by block write would be a way to go now (not a breaking change, current users who are affected by this slow speed would use it) and leaving byte write/read as they are for now (note can be added to the docs if we havent yet, something along the lines - byte writes in the loops might be slow due to locking, use block write instead).

+1 for all this work !