@JanneKiiskila 2017-05-04
Yep, can this make mbedOS 5.5 release? We should try to.