@YarivCol 2017-05-12
you should also call acquire in the method lock, i have a bug (weird signal while switching owner of spi)
this bug is caused from selecting device(cs is 0) and than acquiring the spi bus this result in weird signal(change of spi format) seen by the device to solve this issue i suggest to call acquire in the method lock to insure that acquire of the spi bus will happen before device is selected ( cs =0)