@kjbracey-arm 2017-05-15
Over in atmel-rf-driver, we're having problems with too-slow SPI, and in our case we have already disabled the locks via override, but we're still suffering from the time to setup the hardware on each SPI call to the ST driver.

From that, it appears we do require block read/write, but actually extended into the HAL itself, to avoid programming lots of SPI peripheral registers per byte transferred.

The extension could be signalled by adding a DEVICE_SPI_BLOCK_WRITE, and giving SPI.cpp a compatibility fallback via spi_master_write if not provided.

(We'll only be doing 2-byte transfers for register accesses, but still it would be a significant gain over 1-byte transfers).

So my request is that if you're adding that new write method, please provide the hook for a spi_master_block_write().