@0xc0170 2017-05-18
> So my request is that if you're adding that new write method, please provide the hook for a spi_master_block_write().

Do we need another HAL function if we already got 2 that might be sufficient for this (1byte blocking write, block async write ) or not? Would write block method using async HAL spi write function be slow as well?