@kjbracey-arm 2017-05-18
If you can write your block API in terms of the async HAL API, then I guess that's fine. I haven't really dug through to figure out what the potential traps are there, if any. Would it preclude usage from interrupt? (Not that we do that any more).

In our particular case, it turns out all transfers in the time-critical path are only 2 bytes, so if there was significant setup overhead for the async, it might not work out. But for the ST driver it seems that their sync and async HAL calls are very similar (which is indeed why the sync setup is so slow), so maybe it would be equivalent.

As noted on the other discussion, the other pitfall is we need inter-byte gaps during bursts (250ns - about 2 bits). There's no current API to control that, afaict, so we'd probably still need to tweak the HAL.