@0xc0170 2017-05-18
Thanks @kjbracey-arm for the inputs, this should allows us to find a solution to this. I believe async would be too overkill, same for the new HAL function.

> Over in atmel-rf-driver, we're having problems with too-slow SPI, and in our case we have already disabled the locks via override, but we're still suffering from the time to setup the hardware on each SPI call to the ST driver.

This is an issue for ST spi implementation in this case, I assume that other targets do not have this problem or might have. Thus we will have a look at this (we can write simple test to measure time how long it takes for the transmission for targets). The assumption for 1 byte spi write should be - this function should be fairly small/fast and in most cases possible to be inlined (assuming lock done, SPI peripheral configured (we got that SPI* owner there in SPI class)). In case it is slow, hitting this type of problems, the implementation should be looked at. 

We had something similar with GPIO, when some targets have drivers that takes long times to set gpio output high/low.  @Sissors made a very nice timing GPIO table. We might need something like that for SPI byte write. Based on that table, some targets were improved.